Clarify MediaWikiTestCase::getSchemaOverrides()
I assumed that the 'create', 'drop' and 'alter' fields could be used to
limit the effect of a larger schema update when a test only uses a part
of it; debugging the resulting errors wasted a lot of time. Perhaps this
will help to prevent others from making the same mistake.

Change-Id: Iffa661ff0d967b593897667a081ee6e1399e4701