jquery.suggestions: Do not show suggestions on prefilled values
Show the suggestions on load only if the current value differs to the
defaultValue from the HTML. This prevents suggestions if there is
already a prefilled value in the input field.

Also direct call $.suggestions.update() instead of trigger a keypress.

This change is a follow-up to 72f61f7a5930cf03d4e8ddde62c2ef627b05dd69.

Bug: T224524
Change-Id: I501596996a20c62d1497bf66e23c7858b58bf4ea