jquery.suggestions: Load initial suggestions only when focused
Firefox restores the value of the search field on browser history back
on some special pages with OOUI search suggestions. In this case the
value of the search field gets also changed before JavaScript gets
initialized and the search suggestions get loaded.

This change loads the initial suggestions only when the search field is
focused and the value is changed before initializing JavaScript.

Bug: T224952
Change-Id: I30d334ae61b40666f14c2b6f82fe7f66ebc8fba0