Check for valid user info before calling Linker::userLink() in Linker::revUserTools()
If an actor table row has neither an actor_user nor an actor_name
value (which should not have happened, but did), we would call
Linker::revUserTools() with no valid user info. This generated
log noise and displayed bad output. Instead, check for valid user
info, and fall back to a default message if necessary. T225469
exists for correcting the underlying bad data, but adding this
check is still prudent.

Bug: T224368
Change-Id: Ib9816d8b87b995f4fe465004acd608225728e391