Removes updateIfHeightChanged override on rcfilters
This override causes the rcfilter dropdown menu to be incorrectly
positioned when the height of the "tag box" changes,
i.e. when a tag is added and an additional row is created.

This override was originally added as part of T198399 in commit 
a9fc286, but given bb15314 prevents the resize event from firing on 
pageload, that change should mitigate the performance impact of 
this function. 

Bug: T225499
Change-Id: I4a296c7d3c277d5415144aee6434b63a5035d275