Clarify WikiRevision::downloadSource() and related code
UploadImporter was the name of the new class in the first patch sets of
change I4a6c573fc0, but by the time it was merged as commit 25fbfd9b9c
it had been split into an UploadRevisionImporter interface with an
ImportableUploadRevisionImporter implementation; however, not all the
references to those methods were fully updated.

Also, downloadSource() is not actually intended as a replacement method
– it was introduced, already deprecated, only to avoid going through the
deprecation process, with the intention of removing it (and making the
implementation private) once enough time had passed. Update the comments
to clarify this as well.

Change-Id: I50a07c0db102cdc98e5b7c1a176f401b87d627c9