resourceloader: Remove redundant 'group' setting for StartupModule
This feaure exists for controlling cache fragmentation of modules
that are queued for loading. However, the startup module is never
"queued" for loading.

It's loading is hardwired into ResourceLoaderClientHtml and doesn't
use any batching and couldn't logically, given that by design it
has to be the first thing that loads and cannot be combined with
other code (because we need to be able to return early and control
what else gets loaded from there).

[0] https://www.mediawiki.org/wiki/ResourceLoader/Architecture#Balance

Change-Id: If8900c538ce623f6a9a68a3bbef0927b4b6ae346