build: Add missing dashes in .travis.yml (restore PHP 7.1/7.2 tests)
Follows-up f638c599d5, in which I accidentally removed the leading
dashes from the entries in the 'include' array.

Apparently when you do that, it just assumes you intended to
have a dash in front of the first line and create a single-item
object in which you have the same key four times, where the last
one wins. Thus, testing only 1 variant: php 7.0.

Change-Id: I3fd77749017d15cc60daf9a531417b55c604e7ff