Adapt Recent Changes advanced filters for mobile usage
Changes the behaviour of the rcfilter search input by essentially
turning it into a button for mobile devices. Depending on the value
of `OO.ui.isMobile()` the input is set to readonly mode and given
a shorter message and different icon.

Setting the search input to readonly prevents onscreen keyboards from
being actived, but still opens the filter menu, so that mobile users
can still add/remove filters, just without the ability to search through
them.

Styles are also modified to make the search input appear as a button by
overriding the default readonly and placeholder styles.

Bug: T225499, T223230
Change-Id: Iaa67369542e658d3571d957a204daa7a53d1e520