Tablesorter: Use localeCompare
This should reduce the need for the collationRegex.
It also removes the need for the lowercase based sorting, and
lowercase vs. uppercase order is now determined by the selected
locale.

Android Webviews don't support this, so fallback to the old sort
routine is required.

Have not removed collation regex yet, as the
level of support for specific locales is not entirely clear to me.
Maybe something we should gather statistics for.

Bug: T32674
Bug: T47070
Change-Id: Ie7fe7317471eb12d6a65dfe3c68858c7dc0e92dc