rdbms: merge openConnection into getConnection in LoadBalancer
Add error suppression feature to getConnection() itself so that
callers can easily be migrated to it. Deprecate openConnection(),
which is poorly named and had a poor division of responsibility
with getConnection().

Split out some new private methods for readability:
* sanitizeConnectionFlags(), which warns about and strips unusable flags
* enforceConnectionFlags(), which handles some DBO_TRX sanity logic
* lazyLoadReplicationPositions(), which handles the chronology callback

Make sure "waitForPos" is loaded before pickReaderIndex() so that the
optimization of being extra picky about replication lag (in seconds)
actually applies.

Change-Id: I225ed6e6edf39cdf237f2322ea59f35d30f2e01a