Follow-up 9c9cfa2ec3d7: fix non-session entry point error
Even though we shouldn't really need to, we have to set the language on
the RawMessage we generate when the tag message doesn't exist. Because
both MessageLocalizer and RawMessage are kind of hacky and not fully
supported, I had to hack around the various gaps in our infrastructure.
See my comment on the task for details.

Bug: T227233
Change-Id: I5ab3e9b1e8d34b1af531dba33b2f92fcd7f12192