installer: Call getInfoBox() with raw wikitext, not transformed text
This will be run through the parser by getInfoBox(), and is documented
as taking raw wikitext, such as from Message::plain(). Calling with
text() is needlessly slower and may product incorrect results due to
applying transformations twice.

Also fix indentation for a wfMessage() call that was confusing
previously, seemingly passing 5 parameters with $4 being '<downloadlink/>'
when actually it is $3 and the line above continues a previous parameter.

Bug: T227297
Change-Id: Ice432e08b1c5f52db097f7c9f4fe6d69e1b04c2d