Default to qqx in SpecialPageExecutor
It doesn’t make sense to make the tests depend on the surrounding wiki’s
content language, so let’s use qqx unless the test specifies something
different. This also improves performance because translations no longer
have to be loaded.

The idea is stolen from I8ff715ac53, but this is a more aggressive
version that might improve performance for more tests.

Depends-On: Ica7547812716119e0d0ad504cf52b59358d3f0af
Depends-On: I2e36d7fb8eddc06053b4297213ef26df6b40e17d
Change-Id: I67632cc226c9344bfa5c1ea21d596f34f6b23fab