Make LinksUpdate no longer extend EnqueueableDataUpdate
LinksUpdate does not match RefreshLinksJob since the former is only a subset
of the later. Also, DeferredUpdates::doUpdates() only runs in "enqueue" mode
for cases in MediaWiki::restInPeace() if there is no post-send support.

In a future commit, the deferred callback in which LinksUpdate runs
currently, will be abstracted into its own deferred update, which
will then bring back EnqueueableDataUpdate for this update.

Bug: T206283
Change-Id: I0680be445e8b8e8d0dba85df135b84640f4fcb81