Remember checkbox state on Special:Block if checkbox disabled
Many of the blocking options on Special:Block are only valid for
certain inputs. These options are represented with checkboxes. If
an option is not valid with the current form inputs, its checkbox
is disabled (and not submitted) and it is set to false by default.
To make this clear to the user, we uncheck the checkbox when it
is disabled.

This stores the value of the checkbox before it was disabled and
restores it if the checkbox is enabled again.

Bug: T221117
Change-Id: I3e52d7c3eb70f2798de3fb5ac3824d56e73306a6