installer: Remove <doclink/> parser function and last use of it
Pass the url to the Copying page as a message parameter instead.
This results in identical text and styling. Not even "external link"
vs plain link styling difference, because that doesn't exist in the
installer stylesheet.

The message is renamed in this commit because it is a breaking
change (due to <doclink> being removed, and due to $2 being
required).

Bug: T227297
Change-Id: I73efd9ecb193523a7e384a7a40c768d32b2361b8