resourceloader: Add Config parameter to packageFiles callbacks
Callbacks used to generate the contents of virtual files in packageFiles
modules only received a ResourceLoaderContext object. They could access
the Config object through ResourceLoaderContext::getConfig(), but that
method is deprecated. Pass the Config object as a second parameter to
these callbacks, so that they don't have to use a deprecated method.

Change-Id: Ia4666914e9b07f298ee5ae30ae8c70a3e83b0910